Tommy Wiseau
Tommy wisao is an interesting case
he's well built and pretty ripped even nowdays
he has a very unique and instantly recognizable face
he got alot of style to the way he handles himselfs and it always feels genuine
and the kid got balls
but, when it came time to make his own film, he made a few crucial mistakes
one: he made a film about something he did not understand enough about. leaving the whole thing feling like it was made by an alien
second, he casted himself in the wrong role
the problem with his role was not the role himself, it was because he played it.
even if he got a professional full makeover with a facelift and a complete reconstruction and speechwork and everything he just could not fit that role.
some people just don't fit certain roles
like you won't put jon cena unironically for the role of a happy go lucky ballerina
was he to start his acting career 20 years earlier, he would have fit just right in with the crazy, over the top, muscular, foreign, unique actors.
but even with the fact that no one thing in particular is awful about any of these aspects together they ceated the perfect storm of a movie called "the room""
the most infamous movie of all time, for being so bad it's hillarious
Some people stick to just enjoying the film, they get a chuckle out of the horrible things it is and leave satisfied out of a good laugh
but if you dig deeper you find a truly interesting story that can stand as a cautionary tale and provide deeper understanding into the idea of why things fail, and provide the question: is something failing so bad that it placed it's location in history truly bad or is it good?
It's the perfect, 20th century Icarus story, about a man who flew too close to the sun that his wings couldn't handle it.
But it's also important not to take the wrong message out of these stories
instead of thinking: "your wings will burn cause of your own ambition therefore you should not have ambition"
you need to think: "ambition could burn your wings, the wings are the problem, so I need to find a way to make the wings stronger and better built in order for them not to fall apart, and try to fly in a way that my ambition can be as high as it can without burning my wings"
Infact, it not only offers us the solution to Icaruse's problem
it shows us the direct solution to Tommy's dillema
Because we know the film's direct attention to things Tommy was misinformed about. we can understand how switching around just a few of the pieces could have landed this movie in the successful legendaries.
Showing that understanding what you are talking about or having somebody know what's up is extremely important, and sometimes you have to understand and allow other perspectives cause your own might not be trained to understand the problem or the inaccuracy.
The room infact is a very clever premise that touches deeply upon the evil of a clockwork society, without sticking to the cliche, and showing the burden of an avrage joe who unknowingly turned from just another  person, to the provider of other people who unthankfully feed off him and it drives him to insanity
And so, Tommy keeps appearing here, there, and everywhere. Hell, he might actually one day land a role that fits him, but one day, one day we can truely look at the brilliance behind the room and the brilliance behind it's failure and use it to propel ourselves into Excellence.