The list of monsters so far, as well as tally of all veriety and in black and white the 4 bosses which are yet to be done
The beasts and the death spirits are sort of similar and the fiends and the fighters are too, so all in all we have 5 of those, I want to add 3 more building girls and 2 more plant girls and close it off at 20,  one of those has to be a slime girl. then it's down to the 4 bosses and the protag and that should be planty for starters.